+++ spi.c
@@ -1,34 +1,26 @@
 #include "spi.h"
 
 #include <avr/io.h>
-// #include <avr/interrupt.h>
 
 inline void SetChipSelectHigh(SpiDevice *const dev) {
     PORTB |= (1<<dev->chipSelect);
-    // PORTB |= (1<<PB3);
 }
 
 inline void SetChipSelectLow(SpiDevice *const dev) {
     PORTB &= ~(1<<dev->chipSelect);
-    // PORTB &= ~(1<<PB3);
 }
 
 inline void ToggleSerialClock(SpiDevice *const dev) {
     PORTB |= (1<<dev->serialClock);
     PORTB &= ~(1<<dev->serialClock);
-    // PORTB |= (1<<PB2);
-    // PORTB &= ~(1<<PB2);
-
 }
 
 void SpiSendByte(SpiDevice *const dev, uint8_t data) {
-    for (int bit = 7; bit >= 0; --bit) {
+    for (uint8_t bit = 7; bit >= 0; --bit) {
         if ((data & (1 << bit))) {
             PORTB |= (1 << dev->serialDataInput);
-            // PORTB |= (1 << PB1);
         } else {
             PORTB &= ~(1 << dev->serialDataInput);
-            // PORTB &= ~(1 << PB1);
         }
         ToggleSerialClock(dev);
     }
